LPS-196468 Make the search and the empty state of the LeftSidebar work together
