+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/LeftSidebar/LeftSidebarTreeView.tsx
@@ -8,14 +8,10 @@ import {Text, TreeView} from '@clayui/core';
 import {ClayDropDownWithItems} from '@clayui/drop-down';
 import Icon from '@clayui/icon';
 import {ClayTooltipProvider} from '@clayui/tooltip';
-import {
-	API,
-	getLocalizableLabel,
-	stringIncludesQuery,
-} from '@liferay/object-js-components-web';
+import {API, getLocalizableLabel} from '@liferay/object-js-components-web';
 import classNames from 'classnames';
 import {openToast, sub} from 'frontend-js-web';
-import React, {useMemo, useState} from 'react';
+import React from 'react';
 import {
 	FlowElement,
 	Node,
@@ -36,19 +32,20 @@ const TYPES_TO_SYMBOLS = {
 };
 
 export default function LeftSidebarTreeView({
-	query,
-	setEmptySearch,
+	expandedKeys,
+	leftSidebarOtherObjectFoldersItems,
+	leftSidebarSelectedObjectFolderItem,
+	setExpandedKeys,
 	showActions,
 }: {
-	query: string;
-	setEmptySearch: (value: boolean) => void;
+	expandedKeys: Set<React.Key>;
+	leftSidebarOtherObjectFoldersItems: LeftSidebarItem[];
+	leftSidebarSelectedObjectFolderItem: LeftSidebarItem;
+	setExpandedKeys: React.Dispatch<React.SetStateAction<Set<React.Key>>>;
 	showActions?: boolean;
 }) {
-	const [expandedKeys, setExpandedKeys] = useState<Set<React.Key>>(
-		new Set(['uncategorized'])
-	);
 	const [
-		{elements, leftSidebarItems, selectedObjectFolder},
+		{elements, selectedObjectFolder},
 		dispatch,
 	] = useObjectFolderContext();
 	const store = useStore();
@@ -72,38 +69,6 @@ export default function LeftSidebarTreeView({
 			symbol={hiddenObjectDefinitionNode ? 'hidden' : 'view'}
 		/>
 	);
-
-	const filteredLeftSidebarItems = useMemo(() => {
-		
-		const keys = [] as string[];
-		setEmptySearch(false);
-
-		const newLeftSidebarItems = leftSidebarItems.map((leftSidebarItem) => {
-			if (!leftSidebarItem.leftSidebarObjectDefinitionItems) {
-				return leftSidebarItem;
-			}
-
-			const newLeftSidebarObjectDefinitionItems = leftSidebarItem.leftSidebarObjectDefinitionItems.filter(
-				(leftSidebarObjectDefinitionItem) =>
-					stringIncludesQuery(
-						leftSidebarObjectDefinitionItem.label,
-						query
-					)
-			);
-
-			keys.push(leftSidebarItem.name);
-
-			return {
-				...leftSidebarItem,
-				id: leftSidebarItem.name,
-				leftSidebarObjectDefinitionItems: newLeftSidebarObjectDefinitionItems,
-			};
-		});
-
-		setExpandedKeys(new Set(keys));
-
-		return newLeftSidebarItems;
-	}, [leftSidebarItems, query]);
 
 	const handleMove = async ({
 		objectDefinitionId,
@@ -193,28 +158,6 @@ export default function LeftSidebarTreeView({
 		});
 	};
 
-	const leftSidebarOtherObjectFoldersItems = filteredLeftSidebarItems.filter(
-		(filteredLeftSidebarItem) =>
-			filteredLeftSidebarItem.objectFolderName !==
-				selectedObjectFolder.name &&
-			filteredLeftSidebarItem.leftSidebarObjectDefinitionItems?.length !==
-				0
-	);
-
-	leftSidebarOtherObjectFoldersItems.sort((a, b) =>
-		a.objectFolderName > b.objectFolderName
-			? 1
-			: b.objectFolderName > a.objectFolderName
-			? -1
-			: 0
-	);
-
-	const leftSidebarSelectedObjectFolderItem = filteredLeftSidebarItems.find(
-		(filteredLeftSidebarItem) =>
-			filteredLeftSidebarItem.objectFolderName ===
-			selectedObjectFolder.name
-	) as LeftSidebarItem;
-
 	const linkedObjectDefinitions = leftSidebarSelectedObjectFolderItem.leftSidebarObjectDefinitionItems?.filter(
 		(leftSidebarObjectDefinitionItem) =>
 			leftSidebarObjectDefinitionItem.type === 'linkedObjectDefinition'
@@ -246,12 +189,6 @@ export default function LeftSidebarTreeView({
 				leftSidebarObjectDefinitionItems: newLeftSidebarObjectDefinitionItems,
 			};
 		}
-	);
-
-	setEmptySearch(
-		!newLeftSidebarOtherObjectFolderItems.length &&
-			leftSidebarSelectedObjectFolderItem.leftSidebarObjectDefinitionItems
-				?.length === 0
 	);
 
 	return (
